Make goRuntimeEnvVarsMetrics function Go version dependent
